libs/secrets/pyproject.toml
[tool.pytest.ini_options]
addopts = "--html='../reports/secrets/unittests/html/index.html' --junitxml='../reports/secrets/unittests/junit.xml'"

[tool.poetry]
name = "secrets"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = "Proprietary"
readme = "README.md"
group = { }

  [[tool.poetry.packages]]
  include = "src"

  [tool.poetry.dependencies]
  python = "^3.12"
  pika = "^1.3.2"
  google-cloud-secret-manager = "^2.18.3"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
